X neo/protogen: Catch length checks overflows on decode
For example a list is encoded as

	l        u32
	[l]item  itemType

on decode len is read from data stream and for rest of data len(data) is
checked to be < l*sizeof(item).

However since l is u32 and sizeof(item) is just number the result of `l
* sizeof(item)` has also u32 type. However it could overflow e.g. for

	l		= 0x20000000
	sizeof(item)	= 8

with the l*sizeof(item) being = u32(0)	(exactly zero) -> oops.

Avoid the problem by doing all checking arithmetics with u64 ints.

+++ go/neo/proto.go
@@ -244,7 +244,7 @@ func (a *Address) neoEncode(b []byte) int {
 	return n
 }
 
-func (a *Address) neoDecode(b []byte) (uint32, bool) {
+func (a *Address) neoDecode(b []byte) (uint64, bool) {
 	n, ok := string_neoDecode(&a.Host, b)
 	if !ok {
 		return 0, false
@@ -288,7 +288,7 @@ func (t IdTime) neoEncode(b []byte) int {
 	return 8
 }
 
-func (t *IdTime) neoDecode(data []byte) (uint32, bool) {
+func (t *IdTime) neoDecode(data []byte) (uint64, bool) {
 	if len(data) < 8 {
 		return 0, false
 	}
@@ -1055,7 +1055,7 @@ type Truncate struct {
 type customCodec interface {
 	neoEncodedLen() int
 	neoEncode(buf []byte) (nwrote int)
-	neoDecode(data []byte) (nread uint32, ok bool)	// XXX uint32 or int here?
+	neoDecode(data []byte) (nread uint64, ok bool)	// XXX uint64 or int here?
 }
 
 func byte2bool(b byte) bool {
@@ -1104,16 +1104,16 @@ func string_neoEncode(s string, data []byte) int {
 	return 4 + l
 }
 
-func string_neoDecode(sp *string, data []byte) (nread uint32, ok bool) {
+func string_neoDecode(sp *string, data []byte) (nread uint64, ok bool) {
 	if len(data) < 4 {
 		return 0, false
 	}
 	l := binary.BigEndian.Uint32(data)
 	data = data[4:]
-	if uint32(len(data)) < l {
+	if uint64(len(data)) < uint64(l) {
 		return 0, false
 	}
 
 	*sp = string(data[:l])
-	return 4 + l, true
+	return 4 + uint64(l), true
 }
